fix: 完善 upms

+++ zeus-system/zeus-upms/zeus-upms-biz/src/main/java/tech/alexchen/zeus/upms/service/impl/SysMenuServiceImpl.java
@@ -1,9 +1,13 @@
 package tech.alexchen.zeus.upms.service.impl;
 
-import cn.hutool.core.collection.CollUtil;
-import cn.hutool.core.lang.Assert;
+import cn.hutool.core.lang.tree.Tree;
+import cn.hutool.core.lang.tree.TreeUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import tech.alexchen.zeus.common.core.exception.ResponsiveRuntimeException;
+import tech.alexchen.zeus.upms.api.constant.SysConstant;
+import tech.alexchen.zeus.upms.api.constant.UpmsResponseCode;
 import tech.alexchen.zeus.upms.api.dto.SysMenuSaveDTO;
 import tech.alexchen.zeus.upms.api.dto.SysMenuUpdateDTO;
 import tech.alexchen.zeus.upms.convert.SysMenuConverter;
@@ -23,12 +27,15 @@ public class SysMenuServiceImpl implements SysMenuService {
     private final SysMenuMapper mapper;
     private final SysMenuConverter converter;
 
-    private static final Long ROOT_MENU_ID = 0L;
-
     @Override
     public Long saveMenu(SysMenuSaveDTO dto) {
-        checkParentMenuExists(dto.getParentId());
-        checkDuplicateMenuInfo(dto.getName());
+        if (!SysConstant.ROOT_MENU_ID.equals(dto.getParentId()) && mapper.isParentMenuNotExists(dto.getParentId())) {
+            throw new ResponsiveRuntimeException(UpmsResponseCode.SYS_MENU_PARENT_NOT_EXISTS);
+        }
+        // 检查同级别下是否存在相同名称的菜单
+        if (mapper.isDeptNameExists(dto.getName(), dto.getParentId(), null)) {
+            throw new ResponsiveRuntimeException(UpmsResponseCode.SYS_MENU_NAME_DUPLICATE);
+        }
         SysMenu entity = converter.toEntity(dto);
         mapper.insert(entity);
         return entity.getId();
@@ -36,15 +43,24 @@ public class SysMenuServiceImpl implements SysMenuService {
 
     @Override
     public void updateMenu(SysMenuUpdateDTO dto) {
-        checkParentMenuExists(dto.getParentId());
+        if (!SysConstant.ROOT_MENU_ID.equals(dto.getParentId()) && mapper.isParentMenuNotExists(dto.getParentId())) {
+            throw new ResponsiveRuntimeException(UpmsResponseCode.SYS_MENU_PARENT_NOT_EXISTS);
+        }
+        // 检查同级别下是否存在相同名称的菜单
+        if (mapper.isDeptNameExists(dto.getName(), dto.getParentId(), dto.getId())) {
+            throw new ResponsiveRuntimeException(UpmsResponseCode.SYS_MENU_NAME_DUPLICATE);
+        }
         SysMenu entity = converter.toEntity(dto);
         mapper.updateById(entity);
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public void removeMenuById(Long id) {
-        List<SysMenu> menus = mapper.selectMenusByParentId(id);
-        Assert.isTrue(CollUtil.isEmpty(menus), "Cannot delete a menu with submenus");
+        if (!mapper.canMenuDeleted(id)) {
+            throw new ResponsiveRuntimeException(UpmsResponseCode.SYS_MENU_HAS_SUBMENU);
+        }
+        // TODO 从角色和用户关联信息中，删除该菜单
         mapper.deleteById(id);
     }
 
@@ -54,26 +70,21 @@ public class SysMenuServiceImpl implements SysMenuService {
     }
 
     @Override
-    public List<SysMenu> getMenuList() {
-        return mapper.selectList(null);
+    public List<Tree<Long>> getMenuTree(Long parentId) {
+        List<SysMenu> sysMenus = mapper.selectMenusByParentId(parentId);
+        // 构建树型结构
+        return TreeUtil.build(sysMenus, parentId, (menu, tree) -> {
+            tree.setId(menu.getId());
+            tree.setName(menu.getName());
+            tree.setParentId(menu.getParentId());
+            tree.setWeight(menu.getSort());
+            tree.putExtra("sort", menu.getSort());
+            tree.putExtra("type", menu.getType());
+            tree.putExtra("permission", menu.getPermission());
+            tree.putExtra("path", menu.getPath());
+            tree.putExtra("icon", menu.getIcon());
+            tree.putExtra("component", menu.getComponent());
+        });
     }
 
-    /**
-     * 检查父级菜单是否存在
-     */
-    private void checkParentMenuExists(Long parentId) {
-        if (ROOT_MENU_ID.equals(parentId)) {
-            return;
-        }
-        SysMenu parentMenu = mapper.selectById(parentId);
-        Assert.notNull(parentMenu, "No Menu with id {} exists", parentId);
-    }
-
-    /**
-     * 检查名称是否重复
-     */
-    private void checkDuplicateMenuInfo(String name) {
-        SysMenu menu = mapper.selectByName(name);
-        Assert.isNull(menu, "Menu name {} duplicate", name);
-    }
 }
